# Special-case 1 module per assembly in Assembly.GetTypes/DefinedTypes